pyproject.toml
[project]
name = "auto-create-email"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
authors = [
    {name = "anti-society-projectx",email = "anti-society-projectx@gmail.com"}
]
readme = "README.md"
dependencies = [
    "aiolimiter>=1.2.1",
    "httpx[socks]>=0.28.1",
    "pyfiglet>=1.0.4",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "termcolor>=3.3.0",
]
